job-tracker-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
httpx
cloudinary
psycopg2-binary
python-dotenv
